requirements.txt
Django>=1.11,<1.11.99
mysqlclient>=1.3.13,<1.3.99
python-ldap==3.2.0
requests==2.20,<2.20.99
django-watchman==0.15.0,<0.15.99
django-registration-redux==1.4
django-extensions==1.7.4
whitenoise==3.3.1,<3.3.99
djangosaml2==0.17.2
django-debug-toolbar>=1.10.1,<1.10.99
xlrd==1.0.0
xlwt==1.2.0

libsass==0.16.0
django-cron>=0.5.1,<0.5.99
future-fstrings>=0.4.4,<0.4.99
python-decouple>=3.1,<3.99

ptvsd>=4.2.0,<4.2.99
django-ptvsd-debug==1.0.3
python-dateutil==2.8.0
